Moved resource display to it's own component
