+++ libs/designer/src/lib/ui/templates/review/ReviewExisting.tsx
@@ -2,10 +2,7 @@ import type { RootState } from '../../../core/state/templates/store';
 import { useIntl } from 'react-intl';
 import { makeStyles, Text, tokens } from '@fluentui/react-components';
 import { useSelector } from 'react-redux';
-import { Spinner, SpinnerSize } from '@fluentui/react';
-import { useSubscriptions } from '../../../core/templates/utils/queries';
-import { useMemo } from 'react';
-import { useTemplatesStrings } from '../templatesStrings';
+import { ResourceDisplay } from './ResourceDisplay';
 
 const useStyles = makeStyles({
   actionName: {
@@ -21,7 +18,7 @@ type ReviewExistingProps = {
 
 export const ReviewExisting = ({ resourceOverrides }: ReviewExistingProps) => {
   const intl = useIntl();
-  const { existingWorkflowName, subscriptionId, location, resourceGroup } = useSelector((state: RootState) => state.workflow);
+  const { existingWorkflowName } = useSelector((state: RootState) => state.workflow);
   const { enableResourceSelection } = useSelector((state: RootState) => state.templateOptions);
 
   const intlText = {
@@ -31,14 +28,6 @@ export const ReviewExisting = ({ resourceOverrides }: ReviewExistingProps) => {
       description: 'Label for workflow name',
     }),
   };
-
-  const { resourceStrings } = useTemplatesStrings();
-
-  const { data: subscriptions, isLoading: subscriptionLoading } = useSubscriptions();
-  const subscriptionDisplayName = useMemo(
-    () => subscriptions?.find((sub) => sub.id === `/subscriptions/${subscriptionId}`)?.displayName ?? '-',
-    [subscriptions, subscriptionId]
-  );
 
   const styles = useStyles();
 
@@ -51,22 +40,7 @@ export const ReviewExisting = ({ resourceOverrides }: ReviewExistingProps) => {
         </Text>
       </div>
 
-      {enableResourceSelection && (
-        <div className="msla-templates-review-block basics">
-          <div className="msla-templates-review-block">
-            <Text>{resourceStrings.SUBSCRIPTION}</Text>
-            {subscriptionLoading ? <Spinner size={SpinnerSize.xSmall} /> : <Text weight="semibold">{subscriptionDisplayName}</Text>}
-          </div>
-          <div className="msla-templates-review-block">
-            <Text>{resourceStrings.LOCATION}</Text>
-            <Text weight="semibold">{location}</Text>
-          </div>
-          <div className="msla-templates-review-block">
-            <Text>{resourceStrings.RESOURCE_GROUP}</Text>
-            <Text weight="semibold">{resourceGroup}</Text>
-          </div>
-        </div>
-      )}
+      {enableResourceSelection && <ResourceDisplay />}
     </div>
   );
 };
